# Optimize icon lookup