Benchmarks for `jgmenu pmenu` without a cache:

Remove cache:

```
rm -rf ~/.local/share/icons/jgmenu-cache
```

Old version:

```
git checkout 19fb4f0
make clean ; make ; make install
# flush disk caches
sudo sh -c 'echo 3 > /proc/sys/vm/drop_caches; for d in /dev/sd* ; do hdparm -F $d ; done'
time jgmenu_run pmenu --die-when-loaded
warning: jgmenu icon-cache has not been created
Failed to load XPM file: /opt/eclipse/icon.xpm
jgmenu_run pmenu --die-when-loaded  0.28s user 1.16s system 82% cpu 1.753 total
# number of disk accesses:
strace -f jgmenu_run pmenu --die-when-loaded 2>&1 | grep -E 'open|stat|access|link' | wc -l
57141
```

New version:

```
git checkout 79ce8b0
make clean ; make ; make install
# flush disk caches
sudo sh -c 'echo 3 > /proc/sys/vm/drop_caches; for d in /dev/sd* ; do hdparm -F $d ; done'
time jgmenu_run pmenu --die-when-loaded
warning: jgmenu icon-cache has not been created
Icons loaded in 0.053772 seconds
Root menu icons loaded
Failed to load XPM file: /opt/eclipse/icon.xpm
Icons loaded in 0.275449 seconds
jgmenu_run pmenu --die-when-loaded  0.28s user 0.06s system 56% cpu 0.605 total
# number of disk accesses:
strace -f jgmenu_run pmenu --die-when-loaded 2>&1 | grep -E 'open|stat|access|link' | wc -l 
3868
```

Note that the loading time is 3x smaller and the number of disk accesses is 15x smaller.

`jgmenu_run cache` will have to be fixed, it is looking up icons one by one:

Old version:

```
# number of disk accesses:
strace -f jgmenu_run cache 2>&1 | grep -E 'open|stat|access|link' | wc -l                  
1384368
```

New version:
```
# number of disk accesses:
strace -f jgmenu_run cache 2>&1 | grep -E 'open|stat|access|link' | wc -l
66826
```

Note that this is much worse than starting without a cache at all! (However, after the cache is created, jgmenu runs quickly. My point is that creating the cache is a bit slow.)